[WEB-110] refactored the profile and requirements edit form
