+++ src/components/Dashboard/EditForm/index.ts
@@ -1,19 +1,23 @@
 import EditButton from "./EditButton";
-import EditProfileForm from "./EditProfileForm";
 import EditInputField from "./EditInputField";
 import EditFormModal from "./EditFormModal";
 import EditSelectField from "./EditSelectField";
 import FormButton from "./FormButton";
 import FormButtonsSection from "./FormButtonsSection";
 import EditTextAreaField from "./EditTextAreaField";
+import EditFormLayout from "./EditFormLayout";
+import ProfileFormFields from "./ProfileFormFields";
+import RequirementFormFields from "./RequirementFormFields";
 
 export {
   FormButton,
   EditButton,
   EditFormModal,
-  EditProfileForm,
   EditInputField,
   EditSelectField,
   FormButtonsSection,
   EditTextAreaField,
+  EditFormLayout,
+  ProfileFormFields,
+  RequirementFormFields,
 };
